requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
billiard==4.2.2
boto3==1.40.44
botocore==1.40.44
bs4==0.0.2
build==1.3.0
cachetools==6.2.0
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chromadb==1.1.0
clerk==0.1.0
clerk-backend-api==1.6.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
composio-core==0.7.4
composio-langchain==0.7.4
contourpy==1.3.2
cryptography==43.0.3
cycler==0.12.1
distro==1.9.0
duckduckgo-search==8.1.1
durationpy==0.10
ecdsa==0.19.1
eval-type-backport==0.2.2
exceptiongroup==1.3.0
faiss-cpu==1.12.0
fastapi==0.118.0
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.32.0
grpcio==1.75.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
inflection==0.5.1
invoke==2.2.0
jiter==0.11.0
jmespath==1.0.1
jose==1.0.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kombu==5.5.4
kubernetes==33.1.0
langchain==0.3.17
langchain-core==0.3.60
langchain-google-genai==2.0.7
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langsmith==0.3.45
lxml==6.0.2
markdown==3.9
markdown-it-py==4.0.0
masonite==0.2.9.4
matplotlib==3.10.6
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
mypy-extensions==1.1.0
narwhals==2.6.0
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.0
openai==1.109.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.3.3
paramiko==4.0.0
pillow==11.3.0
plotly==6.3.1
posthog==5.4.0
primp==0.15.0
prompt-toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.10.6
pydantic-core==2.27.2
pydub==0.24.1
pygments==2.19.2
pyjwt==2.10.1
pynacl==1.6.0
pyparsing==3.2.5
pypdf2==3.0.1
pyperclip==1.11.0
pypika==0.48.9
pyproject-hooks==1.2.0
pysher==1.0.8
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
razorpay==2.0.0
redis==6.4.0
referencing==0.36.2
regex==2025.9.18
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.27.1
rsa==4.9.1
s3fs==0.4.2
s3transfer==0.14.0
scipy==1.15.3
seaborn==0.13.2
selenium==4.36.0
semver==3.0.4
sentry-sdk==2.39.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
speechrecognition==3.14.3
sqlalchemy==2.0.43
starlette==0.48.0
stripe==13.0.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.1
tomli==2.2.1
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typer==0.19.2
types-requests==2.32.4.20250913
typing-extensions==4.15.0
typing-inspect==0.9.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
validator-py==1.2.5
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
